AG/vt- dev test/python/requirements.txt
tornado==6.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
apscheduler==3.10.4
